rename ApplyHyperspace
